changed e2e tests to use nginx and added tests to quick start e2e test

+++ docker/runner.sh
@@ -1,11 +1,6 @@
 #!/bin/bash
 set -e
-envsubst < /config/main.template > /config/main.conf
-envsubst < /config/user.template > /config/user.conf
-# append the proxies.conf if it exists
-if [ -f /etc/joule/configs/proxies.conf ]; then
-    cat /etc/joule/configs/proxies.conf >> /config/main.conf
-fi
+
 
 # run nginx startup scripts
 /bin/sh 15-local-resolvers.envsh
@@ -15,6 +10,14 @@ fi
 # start nginx
 nginx -g "daemon on;"
 
-# start jouled
-rm -f /tmp/joule/pid # remove any existing pid (if container restarts)
-/usr/local/bin/jouled
+# start jouled unless --nginx-only is passed
+if [ "$1" != "--nginx-only" ]; then
+    envsubst < /config/main.template > /config/main.conf
+    envsubst < /config/user.template > /config/user.conf
+    # append the proxies.conf if it exists
+    if [ -f /etc/joule/configs/proxies.conf ]; then
+        cat /etc/joule/configs/proxies.conf >> /config/main.conf
+    fi
+    rm -f /tmp/joule/pid # remove any existing pid (if container restarts)
+    /usr/local/bin/jouled    # start jouled
+fi
